# openapi.yaml
openapi: 3.0.3
info:
  title: Nobody Logger API
  description: |
    Nobody Logger是一个基于WBS多层级时间管理和日志记录系统的RESTful API。
    
    ## 核心功能
    - WBS多层级任务管理（年→半年→季度→月→周→日）
    - 实时时间记录和统计分析
    - 工作分类和标签系统
    - 智能提醒系统
    - 模板系统
    - 自动报告生成
    - 跨设备数据同步
    
    ## 认证方式
    使用JWT Bearer Token进行认证：
    ```
    Authorization: Bearer <jwt_token>
    ```
  version: 1.0.0
  contact:
    name: Nobody Logger API Support
    email: support@nobody-logger.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.nobody-logger.com/v1
    description: 生产环境
  - url: https://staging-api.nobody-logger.com/v1
    description: 测试环境
  - url: http://localhost:3000/v1
    description: 开发环境
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: 用户注册
      description: 创建新用户账户
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Authentication]
      summary: 用户登录
      description: 用户身份验证并获取访问令牌
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /projects:
    get:
      tags: [Projects]
      summary: 获取项目列表
      parameters:
        - name: is_active
          in: query
          description: 过滤活跃项目
          schema:
            type: boolean
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: 项目列表获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedProjectList'

    post:
      tags: [Projects]
      summary: 创建项目
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      responses:
        '201':
          description: 项目创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Project'
...
  /time-logs:
    get:
      tags: [Time Logs]
      summary: 获取时间日志列表
      parameters:
        - name: task_id
          in: query
          description: 任务ID
          schema:
            type: integer
        - name: start_date
          in: query
          description: 开始日期
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: 结束日期
          schema:
            type: string
            format: date
        - name: is_manual
          in: query
          description: 是否手动记录
          schema:
            type: boolean
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: 时间日志列表获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedTimeLogList'

    post:
      tags: [Time Logs]
      summary: 创建时间记录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeLogCreateRequest'
      responses:
        '201':
          description: 时间记录创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TimeLog'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      description: 页码
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      description: 每页数量
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      description: 排序字段 (前缀 - 表示降序)
      schema:
        type: string
        example: "-created_at"
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: 数据冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "操作成功"
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "RESOURCE_NOT_FOUND"
            message:
              type: string
              example: "资源不存在"
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
...
    Pagination:
      type: object
      properties:
        current_page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 5
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false

    # 枚举类型
    LevelType:
      type: string
      enum: [yearly, half_yearly, quarterly, monthly, weekly, daily]
      description: 任务层级类型
    TaskStatus:
      type: string
      enum: [not_started, in_progress, completed, paused, cancelled]
      description: 任务状态
    Priority:
      type: string
      enum: [low, medium, high, urgent]
      description: 优先级
...
    RegisterRequest:
      type: object
      required: [email, password, username, device_id]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          example: "password123"
        username:
          type: string
          maxLength: 100
          example: "用户名"
        device_id:
          type: string
          example: "device-uuid-string"
    LoginRequest:
      type: object
      required: [email, password, device_id]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        device_id:
          type: string
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expires_at:
          type: string
          format: date-time

    # 用户相关
    User:
      type: object
      properties:
        id:
          type: integer
          example: 123
        email:
          type: string
          format: email
        username:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
        sync_version:
          type: integer
...
    Project:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        description:
          type: string
        color:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        sync_version:
          type: integer
    ProjectCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
        color:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
          default: "#1976d2"
...
    PaginatedProjectList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    WBSTask:
      type: object
      properties:
        id:
          type: integer
        project_id:
          type: integer
        parent_id:
          type: integer
          nullable: true
        wbs_code:
          type: string
        name:
          type: string
        description:
          type: string
        level:
          type: integer
          minimum: 0
          maximum: 5
        level_type:
          $ref: '#/components/schemas/LevelType'
        sort_order:
          type: integer
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
        estimated_hours:
          type: number
          format: float
          nullable: true
        actual_hours:
          type: number
          format: float
        status:
          $ref: '#/components/schemas/TaskStatus'
        progress_percentage:
          type: integer
          minimum: 0
          maximum: 100
        priority:
          $ref: '#/components/schemas/Priority'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        sync_version:
          type: integer
...
    TimeLog:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        task_id:
          type: integer
        description:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
          nullable: true
        duration_seconds:
          type: integer
          nullable: true
        is_manual:
          type: boolean
        log_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        task:
          $ref: '#/components/schemas/WBSTask'
    TimeLogCreateRequest:
      type: object
      required: [task_id, start_time]
      properties:
        task_id:
          type: integer
        description:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        is_manual:
          type: boolean
          default: true
...
    PaginatedTimeLogList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TimeLog'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # 数据同步
...